LayoutTests/platform/gtk/fast/dom/gc-attribute-node-expected.txt

Tests that attribute node wrappers are not prematurely garbage collected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a.prop is "set"
FAIL a.prop should be set (of type string). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
